app/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic[email]
httpx
#DB
sqlalchemy
sqlalchemy_utils
psycopg2-binary
asyncpg
aiosqlite
python-jose
jwt
#test
pytest
pytest-asyncio
python-dotenv
